+++ src/main/java/pro/sky/calculator/service/CalculatorServiceImpl.java
@@ -10,22 +10,22 @@ public class CalculatorServiceImpl implements CalculatorService {
     }
 
     @Override
-    public int plus(int num1, int num2) {
+    public long plus(long num1, long num2) {
         return num1 + num2;
     }
 
     @Override
-    public int minus(int num1, int num2) {
+    public long minus(long num1, long num2) {
         return num1 - num2;
     }
 
     @Override
-    public int multiply(int num1, int num2) {
+    public long multiply(long num1, long num2) {
         return num1 * num2;
     }
 
     @Override
-    public double divide(int num1, int num2) {
+    public double divide(long num1, long num2) {
         if (num2 == 0) {
             throw new IllegalArgumentException();
         }
